requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0
python-dateutil>=2.8.0
pyyaml>=6.0
xlsxwriter>=3.1.0